refactor(web): simplify frontend by extracting shared components and removing unused code
- Extract NPCFormFields shared component from duplicate NPC create/edit pages (~600 lines → ~360 lines shared)
- Extract LiveTimer component (was duplicated in dashboard and session detail)
- Extract SessionStatusBadge component (was copy-pasted across 3 session views)
- Move GAME_SYSTEMS constant to types.ts (was duplicated in campaign create and edit pages)
- Remove unused sheet.tsx UI component (never imported)
- Move shadcn CLI from dependencies to devDependencies
- Net reduction: ~380 lines (-4.7%)
